% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeBOGs.R
\name{computeBOGs}
\alias{computeBOGs}
\title{Extract the BOG from Wilcox DE analysis}
\usage{
computeBOGs(
  sce,
  de_out_attr = "DE_out",
  parallel = FALSE,
  cores = NULL,
  verbose = FALSE
)
}
\arguments{
\item{sce}{The SingleCellExperiment object}

\item{de_out_attr}{The metadata attribute in which to find the Wilcox DE
result from \code{computeWilcoxDE}.}

\item{parallel}{A boolean to define if the BOGs are obtained using parallel
programming or not.}

\item{cores}{The number of cores wanted. It defaults to NULL, if parallel is
TRUE, the default is the maximum number of cores - 2. If it is set, but it
is either more than the maximum number of cores or less than 1, it defaults
to 1 (if parallel is TRUE).}

\item{verbose}{A boolean to define if updates on the progress of the function
are printed or not.}
}
\value{
The SCE with the BOGsList saved in metadata, under \code{BOGsList}, plus a
BOGs membership matrix under rowData in \code{BOG}.
}
\description{
Extract the BOG from Wilcox DE analysis
}
\examples{
set.seed(123)
expM <- matrix(data = sample(seq(0,100,by = 0.1), 15000*1000, replace = TRUE), ncol = 1000)
colnames(expM) <- paste0('Cell-', seq(1, 1000))
rownames(expM) <- paste0('Gene-', seq(1, 15000))
sce <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = expM))
group <- sample(paste0('C', seq(1, 20)), 1000, replace = TRUE)
sce$MemVec <- group
sce <- computeWilcoxDE(sce, ifelse(group == 'C1', 1, 0), saveToSCE = TRUE)
sce <- computeBOGs(sce)
}
