% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeOperationalRepresentations.R
\name{computeOperationalRepresentations}
\alias{computeOperationalRepresentations}
\title{Computes the Operational Representations (Profiles, Signatures and BOGs)}
\usage{
computeOperationalRepresentations(
  sce,
  group_attr,
  expM_attr = "logcounts",
  na.rm = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sce}{The single cell experiment object.}

\item{group_attr}{The colData attribute in which to find the membership
vector or membership matrix.}

\item{expM_attr}{The assays attribute in which to find the expression matrix
to use. Defaults to 'logcounts'.}

\item{na.rm}{A boolean variable to define if, in the mean, the NA should be
considered or not. If TRUE, they are removed before the computation of the
mean.}

\item{verbose}{A boolean. If TRUE, informations about the proceeding of the
function are given.}

\item{...}{Additional parameters for the computeBOGs function.}
}
\value{
The SCE object with Profiles, Signatures and BOGs in the rowData
of the sce (under P, S, and BOG respectively), plus the BOGsList and the
result of the DE analysis in metadata under BOGsList and DE_out respectively.
}
\description{
Computes the Operational Representations (Profiles, Signatures and BOGs)
}
\examples{
set.seed(123)
expM <- matrix(data = sample(seq(0,100,by = 0.1), 15000*1000, replace = TRUE), ncol = 1000)
colnames(expM) <- paste0('Cell-', seq(1, 1000))
rownames(expM) <- paste0('Gene-', seq(1, 15000))
sce <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = expM))
group <- matrix(data = sample(seq(0,100,by = 0.1), 1000*20, replace = TRUE), ncol = 20)
colnames(group) <- paste0('Cluster-', seq(1, 20))
rownames(group) <- colnames(sce)
sce$MemMat <- group
sce <- computeOperationalRepresentations(sce, 'MemMat')
}
