% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeProfiles.R
\name{computeProfiles}
\alias{computeProfiles}
\title{Compute profiles from a SCE object and a membership vector or matrix}
\usage{
computeProfiles(sce, group_attr, expM_attr = "logcounts", na.rm = TRUE)
}
\arguments{
\item{sce}{The SingleCellExperiment object}

\item{group_attr}{The colData attribute in which to find the membership
vector or membership matrix}

\item{expM_attr}{The assays attribute in which to find the expression matrix
to use. Defaults to 'logcounts'.}

\item{na.rm}{A boolean variable to define if, in the mean, the NA should be
considered or not. If TRUE, they are removed before the computation of the
mean.}
}
\value{
The SCE with the profile matrix (in which each column is a
cluster/group defined in the membership vector/matrix and each row is a gene)
saved in rowData under 'P'. The values are the average expression of that
gene in that cluster, and if a membership matrix is given with cells having
multiple values for each cluster, the weighted average is computed based on
the weight in the membership matrix.
}
\description{
Compute profiles from a SCE object and a membership vector or matrix
}
\examples{
set.seed(123)
expM <- matrix(data = sample(seq(0,100,by = 0.1), 15000*1000, replace = TRUE), ncol = 1000)
colnames(expM) <- paste0('Cell-', seq(1, 1000))
rownames(expM) <- paste0('Gene-', seq(1, 15000))
sce <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = expM))
group <- matrix(data = sample(seq(0,100,by = 0.1), 1000*20, replace = TRUE), ncol = 20)
colnames(group) <- paste0('Cluster-', seq(1, 20))
rownames(group) <- colnames(sce)
sce$MemMat <- group
sce <- computeProfiles(sce, 'MemMat')
}
