% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSignatures.R
\name{computeSignatures}
\alias{computeSignatures}
\title{Compute signatures}
\usage{
computeSignatures(sce, profile_attr = "P")
}
\arguments{
\item{sce}{The SingleCellExperiment object}

\item{profile_attr}{The rowData attribute in which the expression profiles are
saved. Defaults to 'P'.}
}
\value{
The SCE with the signatures (a matrix genes by groups with the
signature values) saved in rowData under 'S'.
}
\description{
Compute signatures
}
\examples{
set.seed(123)
expM <- matrix(data = sample(seq(0,100,by = 0.1), 15000*1000, replace = TRUE), ncol = 1000)
colnames(expM) <- paste0('Cell-', seq(1, 1000))
rownames(expM) <- paste0('Gene-', seq(1, 15000))
sce <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = expM))
group <- matrix(data = sample(seq(0,100,by = 0.1), 1000*20, replace = TRUE), ncol = 20)
colnames(group) <- paste0('Cluster-', seq(1, 20))
rownames(group) <- colnames(sce)
sce$MemMat <- group
sce <- computeProfiles(sce, 'MemMat')
sce <- computeSignatures(sce, 'P')
}
