% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeWilcoxDE.R
\name{computeWilcoxDE}
\alias{computeWilcoxDE}
\title{Wilcox differential expression analysis}
\usage{
computeWilcoxDE(
  sce,
  binGroup,
  expM_attr = "logcounts",
  padj_method = stats::p.adjust.methods,
  orderOut = FALSE,
  saveToSCE = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{sce}{The SingleCellExperiment object.}

\item{binGroup}{The groups bins that define the comparisons for the DE.}

\item{expM_attr}{The assays attribute in which to find the expression matrix
to use. Defaults to 'logcounts'.}

\item{padj_method}{The method used to adjust the pvalues. It defaults to BH.}

\item{orderOut}{A boolean that defines if the result has to be ordered based
on the AUC. Defauls to FALSE.}

\item{saveToSCE}{A boolean. If TRUE, it saves the output to metadata, under
'DE_out'. Otherwise, it just returns the output. Defaults to FALSE.}

\item{verbose}{A boolean to define if updates on the progress of the function
are printed or not.}
}
\value{
Either the SCE with the result of presto::wilcoxauc differential
expression analysis saved in metadata, in the 'DE_out' attribute (if
saveToSCE is TRUE). Otherwise, just the output of presto::wilcoxauc.
}
\description{
Wilcox differential expression analysis
}
\examples{
set.seed(123)
expM <- matrix(data = sample(seq(0,100,by = 0.1), 15000*1000, replace = TRUE), ncol = 1000)
colnames(expM) <- paste0('Cell-', seq(1, 1000))
rownames(expM) <- paste0('Gene-', seq(1, 15000))
sce <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = expM))
group <- sample(paste0('C', seq(1, 20)), 1000, replace = TRUE)
sce$MemVec <- group
sce <- computeWilcoxDE(sce, ifelse(group == 'C1', 1, 0))
}
