% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dge_out.R
\name{filter_dge_out}
\alias{filter_dge_out}
\title{Filter the result of Wilcox DE analysis}
\usage{
filter_dge_out(
  degOut,
  cutFrac = 0.25,
  cutFC = 0.05,
  cutPadj = 0.01,
  orderOut = TRUE
)
}
\arguments{
\item{degOut}{The result of Wilcox DE analysis}

\item{cutFrac}{A threshold for the percentage of non-zero fature values}

\item{cutFC}{A threshold for the Fold-Change}

\item{cutPadj}{A threshold for the adjusted pvalue}

\item{orderOut}{A boolean variable to order the result based on the AUC (decreasing=TRUE)}
}
\value{
The result of Wilcox DE analysis with only the genes that satisfy
all thresholds.
}
\description{
Filter the result of Wilcox DE analysis
}
\examples{
set.seed(123)
expM <- matrix(data = sample(seq(0,100,by = 0.1), 15000*1000, replace = TRUE), ncol = 1000)
colnames(expM) <- paste0('Cell-', seq(1, 1000))
rownames(expM) <- paste0('Gene-', seq(1, 15000))
sce <- SingleCellExperiment::SingleCellExperiment(assays = list(logcounts = expM))
group <- sample(paste0('C', seq(1, 20)), 1000, replace = TRUE)
sce$MemVec <- group
degOut <- computeWilcoxDE(sce, ifelse(group == 'C1', 1, 0), saveToSCE = FALSE)
out <- filter_dge_out(degOut, cutPadj = 2)
}
