% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_zero_pvals.R
\name{remove_zero_pvals}
\alias{remove_zero_pvals}
\title{Set zero pvalues to the minimum non-zero pvalue times 0.1}
\usage{
remove_zero_pvals(xVec, use_min_machine = TRUE)
}
\arguments{
\item{xVec}{The vector of pvalues}

\item{use_min_machine}{A boolean. If TRUE, 0s are set to the minimum integer
in the machine. Otherwise, the minimum non zero value of the vector divided
by 10.}
}
\value{
The modified vector of pvalues
}
\description{
Set zero pvalues to the minimum non-zero pvalue times 0.1
}
\examples{
vec <- rep(0.05, 10)
vec <- c(vec, 0)
new_vec <- remove_zero_pvals(vec)
}
